version 0.3.1  (2015.08.28)
Major edits to the ANOVA code.

NOTES:

1. Newly supported ANOVA models:
— anova3rm  (three-way design with repeated-measures on all three factors)
- anova2onerm  (now supports unbalanced designs:  i.e. different numbers of subjects for each level of factor A)


2. Non-sphericity corrections not supported in the MATLAB version of spm1d.


3. Repeated-measures ANOVA:
IF (a) the data are 1D and (b) there is only one observation per subject and per condition...
THEN inference is approximate, based on approximated residuals.
TO AVOID THIS PROBLEM:  use multiple observations per subject per condition, and the same number of observations across all subjects and conditions.