Version M.0.4.2 (2016.12.15)
Fixed a bug in “get_test_stats” for two- and three-way ANOVA which raised an error for 1D data.

Example:
spmlist = spm1d.stats.anova2(Y, A, B);
fstats  = spmlist.get_test_stats();

This call to get_test_stats previously produced an error.  Now it returns an (n x Q) array, where n is the number of effects in the ANOVA model, and Q is the number of continuum nodes.